hitrace老化机制优化

+++ frameworks/trace_dump_executor/trace_dump_executor.cpp
@@ -399,7 +399,7 @@ bool TraceDumpExecutor::StartDumpTraceLoop(const TraceDumpParam& param)
             std::lock_guard<std::mutex> lck(traceFileMutex_);
             struct stat info;
             if (stat(traceFile.c_str(), &info) == 0) {
-                loopTraceFiles_.emplace_back(traceFile, 0, info.st_size / BYTE_PER_KB, true);
+                loopTraceFiles_.emplace_back(traceFile, 0, static_cast<uint64_t>(info.st_size / BYTE_PER_KB), true);
             } else {
                 loopTraceFiles_.emplace_back(traceFile);
             }
